fastapi_app/requirement.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
Django==5.2.1
django-tailwind==4.0.1
dnspython==2.7.0
docx2pdf==0.1.8
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cache2==0.2.2
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.7.1
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pendulum==3.1.0
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.13.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
pytz==2025.2
redis==6.4.0
requests==2.32.4
sendgrid==6.12.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
